refactor: lib for download to json, separate component for delete button
