+++ src/support/downloadFile.js
@@ -1,0 +1,11 @@
+export const downloadFile = ({ data, fileName, fileType }) => {
+  // Create a blob with the data we want to download as a file
+  const blob = new Blob([data], { type: fileType })
+  // Create an anchor element and dispatch a click event on it
+  // to trigger a download
+  const a = document.createElement('a')
+  a.download = fileName
+  a.href = window.URL.createObjectURL(blob)
+  a.click()
+  a.remove()
+}
